+++ DearPyGui/dearpygui/dearpygui.py
@@ -1608,7 +1608,7 @@ def font_registry(*, label: str =None, user_data: Any =None, use_internal_label:
 		internal_dpg.pop_container_stack()
 
 @contextmanager
-def group(*, label: str =None, user_data: Any =None, use_internal_label: bool =True, tag: Union[int, str] =0, width: int =0, indent: int =-1, parent: Union[int, str] =0, before: Union[int, str] =0, payload_type: str ='$$DPG_PAYLOAD', drag_callback: Callable =None, drop_callback: Callable =None, show: bool =True, pos: Union[List[int], Tuple[int]] =[], filter_key: str ='', delay_search: bool =False, tracked: bool =False, track_offset: float =0.5, horizontal: bool =False, horizontal_spacing: float =-1) -> Union[int, str]:
+def group(*, label: str =None, user_data: Any =None, use_internal_label: bool =True, tag: Union[int, str] =0, width: int =0, indent: int =-1, parent: Union[int, str] =0, before: Union[int, str] =0, payload_type: str ='$$DPG_PAYLOAD', drag_callback: Callable =None, drop_callback: Callable =None, show: bool =True, pos: Union[List[int], Tuple[int]] =[], filter_key: str ='', delay_search: bool =False, tracked: bool =False, track_offset: float =0.5, horizontal: bool =False, horizontal_spacing: float =-1, xoffset: float =0.0) -> Union[int, str]:
 	"""	Creates a group that other widgets can belong to. The group allows item commands to be issued for all of its members. Must be closed with the end command.
 
 	Args:
@@ -1631,11 +1631,12 @@ def group(*, label: str =None, user_data: Any =None, use_internal_label: bool =T
 		track_offset (float, optional): 0.0f:top, 0.5f:center, 1.0f:bottom
 		horizontal (bool, optional): Forces child widgets to be added in a horizontal layout.
 		horizontal_spacing (float, optional): Spacing for the horizontal layout.
+		xoffset (float, optional): Offset from containing window x item location within group.
 	Yields:
 		Union[int, str]
 	"""
 	try:
-		widget = internal_dpg.add_group(label=label, user_data=user_data, use_internal_label=use_internal_label, tag=tag, width=width, indent=indent, parent=parent, before=before, payload_type=payload_type, drag_callback=drag_callback, drop_callback=drop_callback, show=show, pos=pos, filter_key=filter_key, delay_search=delay_search, tracked=tracked, track_offset=track_offset, horizontal=horizontal, horizontal_spacing=horizontal_spacing)
+		widget = internal_dpg.add_group(label=label, user_data=user_data, use_internal_label=use_internal_label, tag=tag, width=width, indent=indent, parent=parent, before=before, payload_type=payload_type, drag_callback=drag_callback, drop_callback=drop_callback, show=show, pos=pos, filter_key=filter_key, delay_search=delay_search, tracked=tracked, track_offset=track_offset, horizontal=horizontal, horizontal_spacing=horizontal_spacing, xoffset=xoffset)
 		internal_dpg.push_container_stack(widget)
 		yield widget
 	finally:
@@ -3586,7 +3587,7 @@ def add_font_registry(*, label: str =None, user_data: Any =None, use_internal_la
 
 	return internal_dpg.add_font_registry(label=label, user_data=user_data, use_internal_label=use_internal_label, tag=tag, show=show)
 
-def add_group(*, label: str =None, user_data: Any =None, use_internal_label: bool =True, tag: Union[int, str] =0, width: int =0, indent: int =-1, parent: Union[int, str] =0, before: Union[int, str] =0, payload_type: str ='$$DPG_PAYLOAD', drag_callback: Callable =None, drop_callback: Callable =None, show: bool =True, pos: Union[List[int], Tuple[int]] =[], filter_key: str ='', delay_search: bool =False, tracked: bool =False, track_offset: float =0.5, horizontal: bool =False, horizontal_spacing: float =-1) -> Union[int, str]:
+def add_group(*, label: str =None, user_data: Any =None, use_internal_label: bool =True, tag: Union[int, str] =0, width: int =0, indent: int =-1, parent: Union[int, str] =0, before: Union[int, str] =0, payload_type: str ='$$DPG_PAYLOAD', drag_callback: Callable =None, drop_callback: Callable =None, show: bool =True, pos: Union[List[int], Tuple[int]] =[], filter_key: str ='', delay_search: bool =False, tracked: bool =False, track_offset: float =0.5, horizontal: bool =False, horizontal_spacing: float =-1, xoffset: float =0.0) -> Union[int, str]:
 	"""	Creates a group that other widgets can belong to. The group allows item commands to be issued for all of its members. Must be closed with the end command.
 
 	Args:
@@ -3609,11 +3610,12 @@ def add_group(*, label: str =None, user_data: Any =None, use_internal_label: boo
 		track_offset (float, optional): 0.0f:top, 0.5f:center, 1.0f:bottom
 		horizontal (bool, optional): Forces child widgets to be added in a horizontal layout.
 		horizontal_spacing (float, optional): Spacing for the horizontal layout.
+		xoffset (float, optional): Offset from containing window x item location within group.
 	Returns:
 		Union[int, str]
 	"""
 
-	return internal_dpg.add_group(label=label, user_data=user_data, use_internal_label=use_internal_label, tag=tag, width=width, indent=indent, parent=parent, before=before, payload_type=payload_type, drag_callback=drag_callback, drop_callback=drop_callback, show=show, pos=pos, filter_key=filter_key, delay_search=delay_search, tracked=tracked, track_offset=track_offset, horizontal=horizontal, horizontal_spacing=horizontal_spacing)
+	return internal_dpg.add_group(label=label, user_data=user_data, use_internal_label=use_internal_label, tag=tag, width=width, indent=indent, parent=parent, before=before, payload_type=payload_type, drag_callback=drag_callback, drop_callback=drop_callback, show=show, pos=pos, filter_key=filter_key, delay_search=delay_search, tracked=tracked, track_offset=track_offset, horizontal=horizontal, horizontal_spacing=horizontal_spacing, xoffset=xoffset)
 
 def add_handler_registry(*, label: str =None, user_data: Any =None, use_internal_label: bool =True, tag: Union[int, str] =0, show: bool =True) -> Union[int, str]:
 	"""	Adds a handler registry.
@@ -4950,25 +4952,6 @@ def add_raw_texture(width : int, height : int, default_value : Union[List[float]
 	"""
 
 	return internal_dpg.add_raw_texture(width, height, default_value, label=label, user_data=user_data, use_internal_label=use_internal_label, tag=tag, format=format, parent=parent)
-
-def add_same_line(*, label: str =None, user_data: Any =None, use_internal_label: bool =True, tag: Union[int, str] =0, parent: Union[int, str] =0, before: Union[int, str] =0, show: bool =True, xoffset: float =0.0, spacing: float =-1.0) -> Union[int, str]:
-	"""	Places a widget on the same line as the previous widget. Can also be used for horizontal spacing.
-
-	Args:
-		label (str, optional): Overrides 'name' as label.
-		user_data (Any, optional): User data for callbacks.
-		use_internal_label (bool, optional): Use generated internal label instead of user specified (appends ### uuid).
-		tag (Union[int, str], optional): Unique id used to programmatically refer to the item.If label is unused this will be the label.
-		parent (Union[int, str], optional): Parent to add this item to. (runtime adding)
-		before (Union[int, str], optional): This item will be displayed before the specified item in the parent.
-		show (bool, optional): Attempt to render widget.
-		xoffset (float, optional): Offset from containing window.
-		spacing (float, optional): Offset from previous widget.
-	Returns:
-		Union[int, str]
-	"""
-
-	return internal_dpg.add_same_line(label=label, user_data=user_data, use_internal_label=use_internal_label, tag=tag, parent=parent, before=before, show=show, xoffset=xoffset, spacing=spacing)
 
 def add_scatter_series(x : Union[List[float], Tuple[float]], y : Union[List[float], Tuple[float]], *, label: str =None, user_data: Any =None, use_internal_label: bool =True, tag: Union[int, str] =0, parent: Union[int, str] =0, before: Union[int, str] =0, source: Union[int, str] =0, show: bool =True) -> Union[int, str]:
 	"""	Adds a scatter series to a plot.
@@ -7734,7 +7717,6 @@ mvKey_Open_Brace=internal_dpg.mvKey_Open_Brace
 mvKey_Backslash=internal_dpg.mvKey_Backslash
 mvKey_Close_Brace=internal_dpg.mvKey_Close_Brace
 mvKey_Quote=internal_dpg.mvKey_Quote
-mvSameLine=internal_dpg.mvSameLine
 mvInputText=internal_dpg.mvInputText
 mvButton=internal_dpg.mvButton
 mvTool_About=internal_dpg.mvTool_About
@@ -7764,8 +7746,8 @@ mvMenuBar=internal_dpg.mvMenuBar
 mvViewportMenuBar=internal_dpg.mvViewportMenuBar
 mvMenu=internal_dpg.mvMenu
 mvMenuItem=internal_dpg.mvMenuItem
-mvGroup=internal_dpg.mvGroup
 mvChild=internal_dpg.mvChild
+mvGroup=internal_dpg.mvGroup
 mvSliderFloat=internal_dpg.mvSliderFloat
 mvSliderInt=internal_dpg.mvSliderInt
 mvFilterSet=internal_dpg.mvFilterSet